# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Easy SKS API
  description: Open-source flashcard study app for the Sportkuestenschifferschein
    (SKS)
  version: 1.0.0
paths:
...
  /study/due:
    get:
      tags:
      - Study
      summary: Get Due Cards
      operationId: get_due_cards_study_due_get
      parameters:
      - name: topic
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Topic
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudyCardOut'
                title: Response Get Due Cards Study Due Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /study/practice:
    get:
      tags:
      - Study
      summary: Get Practice Cards
      operationId: get_practice_cards_study_practice_get
      parameters:
      - name: topic
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Topic
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudyCardOut'
                title: Response Get Practice Cards Study Practice Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /study/review:
    post:
      tags:
      - Study
      summary: Review Card
      operationId: review_card_study_review_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyCardOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /study/evaluate-answer:
    post:
      tags:
      - Study
      summary: Evaluate Answer
      operationId: evaluate_answer_study_evaluate_answer_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluateAnswerIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluateAnswerOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /ai/transcribe-audio:
    post:
      tags:
      - AI
      summary: Transcribe Audio
      operationId: transcribe_audio_ai_transcribe_audio_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_transcribe_audio_ai_transcribe_audio_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudioTranscriptionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AudioTranscriptionOut:
      properties:
        transcript:
          type: string
          title: Transcript
      type: object
      required:
      - transcript
      title: AudioTranscriptionOut
    Body_transcribe_audio_ai_transcribe_audio_post:
      properties:
        audio:
          type: string
          format: binary
          title: Audio
        language:
          type: string
          title: Language
          default: de
      type: object
      required:
      - audio
      title: Body_transcribe_audio_ai_transcribe_audio_post
    CardContentOut:
      properties:
        text:
          type: string
          title: Text
        images:
          items:
            $ref: '#/components/schemas/CardImageOut'
          type: array
          title: Images
      type: object
      required:
      - text
      - images
      title: CardContentOut
    CardImageOut:
      properties:
        image_id:
          type: string
          title: Image Id
        storage_key:
          type: string
          title: Storage Key
        alt_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Alt Text
      type: object
      required:
      - image_id
      - storage_key
      title: CardImageOut
...
    EvaluateAnswerIn:
      properties:
        card_id:
          type: string
          title: Card Id
        user_answer:
          type: string
          title: User Answer
      type: object
      required:
      - card_id
      - user_answer
      title: EvaluateAnswerIn
    EvaluateAnswerOut:
      properties:
        card_id:
          type: string
          title: Card Id
        awarded_points:
          type: number
          title: Awarded Points
        max_points:
          type: number
          title: Max Points
        verdict:
          type: string
          title: Verdict
        reasoning_summary:
          type: string
          title: Reasoning Summary
        mistakes:
          items:
            type: string
          type: array
          title: Mistakes
        missing_points:
          items:
            type: string
          type: array
          title: Missing Points
        improved_answer_suggestion:
          type: string
          title: Improved Answer Suggestion
        suggested_rating:
          type: integer
          title: Suggested Rating
      type: object
      required:
      - card_id
      - awarded_points
      - max_points
      - verdict
      - reasoning_summary
      - mistakes
      - missing_points
      - improved_answer_suggestion
      - suggested_rating
      title: EvaluateAnswerOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReviewIn:
      properties:
        card_id:
          type: string
          title: Card Id
        rating:
          type: integer
          title: Rating
      type: object
      required:
      - card_id
      - rating
      title: ReviewIn
...
    SchedulingInfoOut:
      properties:
        state:
          type: string
          title: State
        stability:
          type: number
          title: Stability
        difficulty:
          type: number
          title: Difficulty
        reps:
          type: integer
          title: Reps
        lapses:
          type: integer
          title: Lapses
        due:
          type: string
          title: Due
        last_review:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Review
      type: object
      required:
      - state
      - stability
      - difficulty
      - reps
      - lapses
      - due
      title: SchedulingInfoOut
...
    StudyCardOut:
      properties:
        card:
          $ref: '#/components/schemas/study__controller__study_controller__CardOut'
        scheduling_info:
          $ref: '#/components/schemas/SchedulingInfoOut'
      type: object
      required:
      - card
      - scheduling_info
      title: StudyCardOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    study__controller__study_controller__CardOut:
      properties:
        card_id:
          type: string
          title: Card Id
        front:
          $ref: '#/components/schemas/CardContentOut'
        answer:
          $ref: '#/components/schemas/CardContentOut'
        short_answer:
          items:
            type: string
          type: array
          title: Short Answer
        tags:
          items:
            type: string
          type: array
          title: Tags
        exam_sheets:
          items:
            type: integer
          type: array
          title: Exam Sheets
      type: object
      required:
      - card_id
      - front
      - answer
      - short_answer
      - tags
      - exam_sheets
      title: CardOut